Create mfs.c
Became familiar with file allocation tables, endianness, as well as file access. Implemented a user space shell application that is capable of interpreting a FAT32 file system image. The utility did not corrupt the file system image and is robust. No existing kernel code or any other FAT 32 utility code was used in the program.